Changed operator API to notify subscribers on every status change.
Prior to this change, the master would only send `TaskUpdated`
messages to subscribers when the latest known task state on the
agent changed.

This implied that schedulers could not reliably wait for the status
information corresponding to specific state updates, i.e.,
`TASK_RUNNING`, since there is no guarantee that subscribers get
notified during the time when this status update will be included in
the status field.

After this change, `TaskUpdated` messages are sent whenever the latest
acknowledged state of the task changes.

Review: https://reviews.apache.org/r/67575/