Windows: Ported `OsTest.Environment`.
This required replacing `os::raw::environment()` inline as it does not
exist on Windows. Furthermore, it was also discovered that the
documentation is incorrect, and that sometimes environment variables
start with an equal sign. The fix is to simply look for the first
equal sign excluding the first character, otherwise we end up with
empty keys.

Review: https://reviews.apache.org/r/68531