Reduced the number of jemalloc arenas.
It seems that the default high number of allocation arenas (nproc*4)
combined with the high number of libprocess worker threads, which are
mostly "dormant", causes the allocator request a lot of memory from the
system resulting in high process RSS
(https://jira.twitter.biz/browse/AURORA-12085).

Most of jemalloc parameters, including the number of arenas, cannot be
changed in runtime. So the way we configure it is a bit hacky - through
the global string literal. The other way is using MALLOC_CONF
environment variable, which can still be used as an override.

The new number of arenas was selected pretty much arbitrarily. It helped
reduce the agent's RSS (better than 8, same as 2) and existing
benchmarks didn't reveal any performance hits caused by it.

Review: https://reviewboard.twitter.biz/r/931992