Removed `ROOT` requirements for `AgentResourceProviderConfigApiTest`.
These tests required `ROOT` in order to use `filesystem/linux`
isolation this is not a requirement anymore so we can run the tests in
general. These tests appear to be able to run in parallel as well.

We also changed the `AddConflict` test a bit to make it more robust.

Review: https://reviews.apache.org/r/68757