Fixed stout `FsTest.Used` test.
The `f_bsize` elemement denotes the preferred filesystem block
size, not the unit that `f_blocks` is counting in. We need to
use `f_frsize` for that.

Remove the check for used space on `/dev` since that is not even
portable across different Linux distributions.

Review: https://reviews.apache.org/r/68679/