Use total cluster resources as framework sorter "total".
The role sorter and framework sorters differed in that the role
sorter used the total cluster resources as the "total" in which
the allocation is divided to compute shares, whereas the framework
sorters used the role's allocation as the "total".

Per a discussion on the mailing list:

https://s.apache.org/differing-DRF

This patch makes the two types of sorters use the total cluster
as the "total". While this provides slightly different behavior,
it's unlikely any users were relying on this.

Review: https://reviews.apache.org/r/68833