Added persistent volume support to the `disk/xfs` isolator.
Added persistent volume support to the `disk/xfs` isolator. This
implementation largely tracks the `disk/du` implementation in that
we now keep a map of paths in each container info structure. We now
defer quota clean up to project ID reclaimation time so that we can
use the same mechanism for sandbox and persistent volume paths.

We explicitly exclude mount disks from XFS project quotas, but we still
track them so that we can correctly publish their usage information in
the container `DiskStatistics` message. This means that mount disks are
not required to be XFS filesystems or have project quotas configured.

Review: https://reviews.apache.org/r/68401/