Fixed bug in `verify-reviews` due to mismatched types.
Because Python is not type-safe, we encountered a bug in the code
executed on non-Windows platforms that was expecting `output` to be a
normal Python string instead of a Python byte string (with encoded
content). To fix this, we now always decode the bytes into a string,
so that the logic afterwards only has one type to deal with.

Review: https://reviews.apache.org/r/68826/