Updated verify-reviews.py to use current interpreter in subprocesses.
This changes the command used in `support/verify-reviews.py` when
running `support/apply-reviews.py` as a subprocess. It was previously
`"python"`, which is generally Python 2, and is now `sys.executable`.

That way, if verify-reviews.py is run with Python 3 (as it should),
apply-reviews.py will be run with the same Python 3 interpreter. This
should fix the `ImportError` issues we have recently seen in our CI.

Review: https://reviews.apache.org/r/68951/