Replaced 'verify_address_format' in new CLI with 'sanitize_address'.
In the process we cleaned up alot of the logic to make sure that the
addresses that are passed in are of the format we expect and that what
comes out is a sanitized address in canonical form.

As part of this change, we updated a number of call sites to use this
new function.

Review: https://reviews.apache.org/r/66683/