Changed a benign warning log message in slave.cpp to info.
Currently, `UpdateFrameworkMessage` is broadcasted by the master
to all agents regardless of whether the framework actually exists
on the agent (see: https://bit.ly/2OiPB4F). So ignoring info
update for framework due to missing framework on the agent is not
unexpected. A warning message would false alarm the user. This
patch changes the log to info to reduce noises.

Review: https://reviews.apache.org/r/68984