Removed `used` argument in `AgentProfile` in the allocator benchmark.
Currently, it is not easy to initialize frameworks with used
resources in the fixture because when we specify the
`agentProfile`, no framework has been created yet.

Also, we currently do not have any use case for setting `used`
resources during the initialization. We can revisit this once
it becomes necessary.

Review: https://reviews.apache.org/r/69093