Changed usage documentation for new CLI.
Moved the '[options]' in the usage documentation to where they should be
set when using the CLI.  Since we use docopt to parse the command line,
we use its 'options_first' parameter to force all options to be included
before the arguments to a command. This has always been the case, but
the help menu hasn't reflected this. This commit fixes that.

Review: https://reviews.apache.org/r/69026/