Fixed the XFS project ID labeling to not cross mount points.
If we bind mount a directory on the same filesystem, the fts(3)
directory traversal that recursively assigns XFS project IDs isn't
able to detect when it crosses the mount point. This leads it to
erroneously rewrite the project IDs for persistent volumes if they
are bind mounted into the container by the `filesystem/linux`
isolator. The fix is to test whether we crossed a mount point
whenever we descend into a new directory so that we can avoid
assigning project ID beneath it.

Review: https://reviews.apache.org/r/68999/