Updated Python library to properly import submodules.
The primary change was to add imports in '__init__.py' so that when
doing an 'import mesos' from some code using this module, we are
able to access submodules (e.g. 'mesos.http') as part of this.

In the process of making this change, we ran into the error pointed out
in variant 6 at the following link, for how to properly set the version
of a python module both in the code itself, and in its package
definition via setup.py.

https://packaging.python.org/guides/single-sourcing-package-version/

We decided to change to method 1 from this link to address this problem.

Review: https://reviews.apache.org/r/69047/