Moved import of '../lib' from new CLI bootstrap to pip-requirements.txt.
Previously, we were adding the path to this module when setting up the
CLI virtual environment in the bootstrap file. However, we should have
always been pulling it from 'pip-requirements.txt' in order to remain
consistent. We didn't run into any issues with this before, because this
library was unused until now.

Review: https://reviews.apache.org/r/68971/