Add the output file to the hash on CommandInfo::URI.
The `output_file` field of the `CommandInfo::URI` is not combined
into the hash value. This means that the fetcher does not consider two
messages that differ only by the `output_file` as different. By adding
the `output_file` to the hash it is possible for a task to fetch the
same URI to multiple, distinct ouptuts.

Review: https://reviews.apache.org/r/68586/