Fixed fetcher deadlock with duplicate URIs.
If a fetch request contains duplicate URIs that are not already
in the cache, the fetcher would erroneously expect that some other
fetch process is going to download that cache entry. It will then
wait for a Future that will never complete.

The fix is to track whether the cache entry was created in this
fetch, and in that case to simply allow the duplicate URI. In
the fetcher, we check the cache before downloading so that a URIs
can be fetched to distinct output files without being downloaded
multiple times.

Review: https://reviews.apache.org/r/68587/