Updated bootstrap in new CLI to honor relative path in pip requirements.
Previously, bootstrap would fail if there was a relative path in
pip-requirements.txt, but bootstrap was run from a different directory
than where pip-requirements.txt was located.

We fix this now by 'cding' into the directory where the bootstrap file
(and the pip-requirements.tx file) are located before pip installing.