Used mount(8) to make XFS loop device mounts.
When we added the `filesystem/linux` isolator to the XFS isolation
tests, we broke them for older Linux distributions (specifically
CentOS 6). The `filesystem/linux` isolator needs to remount the
work directory in shared mode using `mount(8)`, but `mount(8)`
requires the mount point it is manipulating to be present in
`mtab(5)`. Since the XFS tests simply issued a `mount(2)` system
call, `mtab(5)` wasn't updated and the remount would fail. Modern
Linux distributions don't see this problem because `/` is mounted
shared by default, and `mtab(5)` is a symlink to `/proc/self/mounts`
(i.e. always in sync).

The straightforward fix is to run `mount(8)` to make the XFS test
fixture mounts, so that `mount(8)` always sees a consistent view of
`mtab(5)`.

Review: https://reviews.apache.org/r/69132/