Add a flag to toggle per-framework metrics.
In clusters with high numbers of frameworks, it can be necessary
to control publishing of per framework metrics. Metrics are still
collected for active frameworks, but will not be pulished if this
flag is set. This allows future extraction of these metrics, e.g.
an API call or logs while keeping the code change simple.

Review: https://reviews.apache.org/r/68956/