Fixed bug in 'execute.cpp' with tty-based tasks and no 'containerInfo'.
Previously, we could only launch tasks using the '--tty' flag if they
had a backing docker image (or some other combination of other flags set
that would cause the task to have a 'containerInfo' created for it).

This commit makes sure that if '--tty' is passed, that a containerInfo
gets created and its TTYInfo field gets populated.

Review: https://reviews.apache.org/r/69103/