Fixed LongLivedDefaultExecutorRestart GC test.
This test was incorrectly restarting the agent actor inside the test.
In the flaky test, the agent actor would be started with an auto-
generated PID (i.e. `slave(1)`, `slave(2)`, etc).  Because of how this
generation works, each PID will be unique.  The executor in the test
would be launched under `slave(1)` but the restarted agent would have
a PID of `slave(2)`.  This meant the executor's reregistration would
fail with '404 Not Found' and the executor would be cleaned up.

The executor cleanup would potentially trigger a TASK_LOST status
update; and if that update is sent prior to ending the test, this
will break some mock expectations and cause the test to fail.

This changes the test to always use the same PID for the agent actor.

Review: https://reviews.apache.org/r/69203