Fixed FetcherTest.DuplicateFileURI on OSX.
This test seems to fail because of a mismatch between the implementation
of std::hash and operator== of the CommandInfo::URI object.  The hash
accounts for all fields except for 'cache'.  The equality operator
accounts for all fields except 'cache' **and** 'output_file'.

This mismatch results in replacing the URI value in the fetcher's
mapping, rather than adding a value.  Why the std::unordered_map
checks the equality operator rather than the hash value is unknown,
but may be an implementation detail of clang's stdlib.

Review: https://reviews.apache.org/r/69205