Mounted devpts with `gid=5` by default.
Some versions of glibc (e.g. 2.17 from CentOS 7) are hard-coded
to expect that terminal devices are owned by the tty group. This
causes containers that allocate TTYs to expect to have to chown
the TTY (see grantpt code in glibc), but it may not be possible
to launch a privileged helper to perform the chown (e.g. because
capabilities have been dropped).

Mounting devpts with `gid=5` is the default in CentOS, Docker,
Fedora and Ubuntu, so this should not cause any compatibility
problems.

Review: https://reviews.apache.org/r/69214/