Simplified newline handling in 'test_exec()' test for new CLI.
The test was previously using '.strip()' to compare the command stdout
and the result we expect. This check was incorrect because it could
happen that the output ended in a bunch of extra whitespace that we
would then strip off unknowningly. By replacing the task command to use
'printf' instead of 'echo' (which artifically inserts an extra newline
in the output), we are able to simplify this assertion and make sure the
output is exactly the same as what we expect.

Review: https://reviews.apache.org/r/69237/