Added collectAuthorizations helper to master.hpp.
Adds the helper function 'collectAuthorizations' to master.hpp. This
function allows for a simple way to collect authorization futures and
only if all supplied futures result in an approved authorization will
the returned future return true.

All identified areas that were formally triggering MESOS-9317 are
being updated to make use of this new helper.

A helper function has been chosen and preferred over copying this
pattern into the areas that needed a fix to allow for an efficient and
complete test coverage.

Additionally we are adding a test validating that new helper.

Review: https://reviews.apache.org/r/69369/