Removed separate automake variables for header sources.
In some cases, we were using separate automake sources variables
to list source and header files, and in other cases, we just added
both to the same variable. Standardize on the latter form since it
makes it easier to see where files are listed.

Review: https://reviews.apache.org/r/69323/