Updated 'Makefile.am' to make new CLI build step more reliable.
We list the files in 'MESOS_CLI_SRCDIR' when building the new CLI.
We were previously using 'find' and then removing the files in
the virtual environment but this is not enough as we now also
use tools like 'tox' which creates files we do not care about.

To filter more while simplifying the build step, we now use 'git
ls-files' to get a list of the files we need when building. Git
is a dependency of Apache Mesos and the files we have in the
repository are the only ones required to build the CLI,
making this solution simpler yet more future-proof.

Review: https://reviews.apache.org/r/69084/