Simplified verify-reviews.py to be more similar to the Python 2 script.
We have switched the Python 2 support scripts to Python 3 a few months
ago. During this transition, 'verify-reviews.py' has been modified,
mainly to work better on Windows: https://reviews.apache.org/r/67505/.

The Python 3 script has then replaced the Python 2 one and we now face
an issue with the script, MESOS-9459. Due to all the changes made
between the Python 2 and Python 3 scripts, it is hard for us to
know when the script started failing and why.

To solve this issue, this commit takes the state of 'verify-reviews.py'
as it was in 74121798f24fca372180b8c4bc00b4df07d46240, applies 2to3,
adds the '--out-file' option, then adds the content of:
* ac766c8bd0fba348399d8eac52f75f62e64db64e
* 43eba285cb5c8622a6be8a5bdad637ed53431e85
* 5b348b6070f0d0403cb69b6a7fa638fc46b7ff49
* 568fcdfd29788d9df89a51ffae7969c2bf0ea173

This gives a clean diff between 74121798f24fca372180b8c4bc00b4df07d46240
and this commit, fixes the issue raised in MESOS-9459, and does not
remove existing features. However, this uses less our common library
developed in the chain that landed while migrating from Python 2 to 3.

This should be followed by a refactoring of 'verify-review.py' and
'common.py' to use our library more while not breaking the script
when using Reviewbot.

Review: https://reviews.apache.org/r/69559/