Tested framework subscription with set but empty ID.
This patch add a test for the case where a framework subscribes with a
set, but empty `FrameworkID`. While this has never been officially
supported, the master code currently contains extensive handling for
this case.

This test can be removed if MESOS-9481 gets implemented.

Review: https://reviews.apache.org/r/69557/