Narrowed interface of `ReadOnlyHandler` members.
Previously, the members of ReadOnlyHandler would take a full
Request as parameter, making it hard for clients to reason
about which parts of the request are used internally, and
even harder to guarantee that behaviour into the future.

This commit changes the interface so only the query
parameters get passed.

Review: https://reviews.apache.org/r/69071/