Avoided reviving on behalf of scheduler after agent reconfiguration.
When agent reconfiguration was enabled in Mesos, the allocator was
also updated to remove all offer filters associated with an agent when
that agent's attributes change. In addition, whenever filters for an
agent are removed, the framework is revived for all roles that it is
registered in.

While this ensures that schedulers will have an opportunity to use
resources on an agent after reconfiguration, modifying the scheduler's
suppression may put the scheduler in an inconsistent state, where it
believes it is suppressed in a particular role when it is not.

This patch eliminates the suppression modification code, while
keeping the code which removes a reconfigured agent's offer filters.

Review: https://reviews.apache.org/r/68134/