Exposed `StorageLocalResourceProviderProcess` for testing purpose.
This patch moves the declaration of the SLRP process into an internal
header file and add a `__call` function, so a follow-up test could use
`FUTURE_DISPATCH` to capture a dispatch on an RPC retry.

To simplify the declarations, it also internalizes `RPCTraits` and
introduce new type aliases, and moves `DEFAULT_CSI_RETRY_BACKOFF_FACTOR`
and `DEFAULT_CSI_RETRY_INTERVAL_MAX` to the new header for testing.

Review: https://reviews.apache.org/r/69827