Fixed scheduler library on multiple SUBSCRIBE requests per connection.
The HTTP scheduler API dictates that on a single connection, the
scheduler may only send a single SUBSCRIBE request. Due to recent
authentication related changes, this contract got broken. This patch
restores the contract and adds a test validating that the library is
enforcing it.

Review: https://reviews.apache.org/r/69839/