Made it possible to checkpoint resources without downgrading them.
For backwards compatibility the agent currently downgrades resources
before checkpointing them.

In order to support providing operation feedback for operations
affecting agent default resources, we're going to start atomically
checkpointing resources and operations in a new file.

Since we're going to checkpoint the resources to a new file that is
not used by old agents, there is no need to downgrade resources before
writing them to that file. This patch updates helper functions to make
that possible.

Review: https://reviews.apache.org/r/69791/