Made SLRP recover node-published volumes after reboot.
If a CSI volume has been node-published before a reboot, SLRP will now
try to bring it back to node-published again. This is important to
perform synchronous persistent volume cleanup for `DESTROY`.

To achieve this, in addition to keeping track of the boot ID when a CSI
volume is node-staged in `VolumeState.vol_ready_boot_id` (formerly
`VolumeState.boot_id`), SLRP now also keeps track of the boot ID when
the volume is node-published. This helps SLRP to better determine if a
volume has been published before reboot.

Review: https://reviews.apache.org/r/69892