Added offer filters to static framework configuration.
This patch adds a new field to `FrameworkInfo` which can be used by
frameworks to configure static, per-role resource requirements. We
currently support specifying minimal allocatable resources which is
interpreted by the default, hierarchical allocator for filtering out
resources not relevant to the framework, and for overriding any globally
configured `min_allocatable_resources` on a per framework, per role
basis.

We use a new message `ResourceQuantities` to specify resource
requirements. This class is modelled after the internal class of the
same name, and we provide a conversion from the API to the internal
type.

Review: https://reviews.apache.org/r/69818/